moved api-gateway our of index
